requirements.txt
Flask==2.2.2
python-docx==0.8.11
openai==0.27.0
